// porkangular/src/app/models/usuario.model.ts
export class UsuarioModel{
    _id: any = '';
    nombre: string = '';
    apellido: string = '';
    email: string = '';
    celular: string = '';
    granja: string = '';
    localizacion: string ='';
    password: string = '';
    codigoMostrar: string = '';
    estado: boolean = true;
    idUsuario: string = '';
    pushUserId: string = '';
}